# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Library API - فلسطين",
    "version": "1.0.0",
    "description": "نظام إدارة مكتبة فلسطيني كامل - REST + JWT + RBAC + GraphQL"
  },
  "servers": [
    { "url": "http://localhost:5000/api/v1", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "librarian", "member"] }
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "author": { "type": "string" },
          "availableCopies": { "type": "integer" },
          "totalCopies": { "type": "integer" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "تسجيل الدخول",
        "operationId": "loginUser",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "تم تسجيل الدخول",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          }
        }
      }
    },
    "/books": {
      "get": {
        "summary": "جلب الكتب",
        "operationId": "getBooks",
        "tags": ["Books"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة الكتب",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Book" }
                }
              }
            }
          },
          "401": { "description": "غير مصرح" },
          "403": { "description": "ممنوع" }
        }
      },
      "post": {
        "summary": "إضافة كتاب جديد",
        "operationId": "addBook",
        "tags": ["Books"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "author": { "type": "string" },
                  "isbn": { "type": "string" },
                  "totalCopies": { "type": "integer" }
                },
                "required": ["title", "author", "totalCopies"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء الكتاب",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Book" }
              }
            }
          }
        }
      }
    }
  }
}